tests/requirements.txt
transformers>=4.35.0
datasets>=2.14.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
accelerate>=0.24.0
evaluate>=0.4.0
jiwer>=3.0.0
tqdm>=4.66.0
pyyaml>=6.0.0

